requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.7.0.1
bcrypt==3.1.6
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
docker==3.7.2
docker-compose==1.24.0
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.6
functools32==3.2.3.post2
idna==2.7
ipaddress==1.0.22
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
numpy==1.16.3
opencv-contrib-python==4.1.0.25
paramiko==2.4.2
Pillow==6.0.0
pyasn1==0.4.5
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
PyNaCl==1.3.0
pyxdg==0.25
PyYAML==3.13
requests==2.20.1
SecretStorage==2.3.1
six==1.12.0
texttable==0.9.1
urllib3==1.24.3
websocket-client==0.56.0
